docs(update): redoing whole readme
As I have given an option for README in different lang as the docs site has, but I feel much much difficult to maintain README file in different lang. As I am very happy that docs are in different lang and always they will. But in case of README, there is not many more I have put in README, so I am taking down the README that are in other lang except the main one.

Yeah the docs will remain in different lang, that is necessary there and I don't have to change the docs every time, so the maintain time there will be much less than here.

So remove those lang README and I have bit redone the README part for more practical review.

Thanks

Co-Authored-By: aayush <177418358+aayushrg7@users.noreply.github.com>