docs(readme): add quick link for use-case
I don’t want to open a PR for this small change. Hope this will be fine.

The Quick Test link should be there. People don’t read docs too much, and the How to Use guide is in the last corner of the homepage. So it’s better to have it directly in the README.